requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt --strip-extras pyproject.toml
#
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   httpx
    #   openai
    #   starlette
boto3==1.42.25
    # via ss (pyproject.toml)
botocore==1.42.25
    # via
    #   boto3
    #   s3transfer
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via
    #   reportlab
    #   requests
click==8.3.1
    # via uvicorn
distro==1.9.0
    # via openai
et-xmlfile==2.0.0
    # via openpyxl
fastapi==0.128.0
    # via ss (pyproject.toml)
googleapis-common-protos==1.72.0
    # via opentelemetry-exporter-otlp-proto-http
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via openai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==8.7.1
    # via opentelemetry-api
jiter==0.12.0
    # via openai
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
lxml==6.0.2
    # via python-docx
numpy==2.4.1
    # via pandas
openai==2.15.0
    # via ss (pyproject.toml)
openpyxl==3.1.5
    # via ss (pyproject.toml)
opentelemetry-api==1.39.1
    # via
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   ss (pyproject.toml)
opentelemetry-exporter-otlp-proto-common==1.39.1
    # via opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-http==1.39.1
    # via ss (pyproject.toml)
opentelemetry-proto==1.39.1
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk==1.39.1
    # via
    #   opentelemetry-exporter-otlp-proto-http
    #   ss (pyproject.toml)
opentelemetry-semantic-conventions==0.60b1
    # via opentelemetry-sdk
pandas==2.3.3
    # via ss (pyproject.toml)
pillow==12.1.0
    # via reportlab
prometheus-client==0.23.1
    # via ss (pyproject.toml)
protobuf==6.33.3
    # via
    #   googleapis-common-protos
    #   opentelemetry-proto
pydantic==2.12.5
    # via
    #   fastapi
    #   openai
    #   ss (pyproject.toml)
pydantic-core==2.41.5
    # via pydantic
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   pandas
python-docx==1.2.0
    # via ss (pyproject.toml)
python-multipart==0.0.21
    # via ss (pyproject.toml)
pytz==2025.2
    # via pandas
reportlab==4.4.7
    # via ss (pyproject.toml)
requests==2.32.5
    # via opentelemetry-exporter-otlp-proto-http
s3transfer==0.16.0
    # via boto3
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via openai
starlette==0.50.0
    # via fastapi
tqdm==4.67.1
    # via openai
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   openai
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydantic
    #   pydantic-core
    #   python-docx
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3
    # via pandas
urllib3==2.6.3
    # via
    #   botocore
    #   requests
uvicorn==0.40.0
    # via ss (pyproject.toml)
xlrd==2.0.2
    # via ss (pyproject.toml)
zipp==3.23.0
    # via importlib-metadata
